Fixed JSR-303 bean validation in MDS
- added missing imports
- added missing depenendency
- added the provider file(copy from hibernate) to both mds and mds
entities, so that it can be read from their classloader
- updated datanucleus.properties with the validation property
- added IT for bean validation
- remove redundant test datanucleus.properties
- fixed an issue with MdsBundleIT sometimes trying to delete DDE

+++ platform/mds/mds/src/test/java/org/motechproject/mds/it/osgi/MdsBundleIT.java
@@ -610,7 +610,9 @@ public class MdsBundleIT extends BasePaxIT {
         getLogger().info("Cleaning up entities");
 
         for (EntityDto entity : entityService.listEntities()) {
-            entityService.deleteEntity(entity.getId());
+            if (!entity.isDDE()) {
+                entityService.deleteEntity(entity.getId());
+            }
         }
     }
